Extract command and pagination dispatch
